+++ src/test/groovy/com/streamr/client/protocol/message_layer/StreamMessageSpec.groovy
@@ -2,6 +2,7 @@ package com.streamr.client.protocol.message_layer
 
 import com.streamr.client.exceptions.EncryptedContentNotParsableException
 import com.streamr.client.protocol.message_layer.StreamMessage.EncryptionType
+import com.streamr.client.testing.StreamMessageExamples
 import spock.lang.Specification
 
 class StreamMessageSpec extends Specification {
